% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competeMicrobiota.R
\name{allSugarScoresForOneGenome}
\alias{allSugarScoresForOneGenome}
\title{Calculate Completeness Of All Sugar Degradation Pathways Within One Genome}
\usage{
allSugarScoresForOneGenome(geneVec, allSugars, ER)
}
\arguments{
\item{geneVec}{A vector of gene names, representing all sugar degradation enzymes
present in one genome of interest.}

\item{allSugars}{A vector of all sugar degradation pathways.}

\item{ER}{A data.frame describing enzymatic reactions with at least 3 columns: "Gene"
for gene encoding an enzyme, "Reaction.EC" for categorization of catalytic reactions,
and "Sugar" for degradation pathway. Column names need to be EXACTLY the same,
case sensitive.}
}
\value{
A vector containing a genome's score for all sugar degradation pathways.
}
\description{
Use the calculateStepsForOneSugar function for all sugar pathways one by one,
and build a named vector for all pathways.
}
\examples{
\dontrun{
 geneVec <- c("rpe", "rpiB", "eno", "fucK")
 ER <- microCompet::EnzymaticReactions
 ED <- microCompet::EnzymeDistribution
 allSugars <- sort(unique(ED$Sugar))
 sugarScoreVec <- allSugarScoresForOneGenome(geneVec, allSugars, ER)
 sugarScoreVec
}

}
