% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competeMicrobiota.R
\name{calculateStepsForOneSugar}
\alias{calculateStepsForOneSugar}
\title{Calculate Number Of Sugar Specific Degradation Enzymes In A Vector}
\usage{
calculateStepsForOneSugar(geneVec, sugar, ER)
}
\arguments{
\item{geneVec}{A vector of gene encoding sugar degradation enzymes.}

\item{sugar}{The desired sugar degradation pathway. A simple sugar such as "fructose".}

\item{ER}{A data.frame describing enzymatic reactions with at least 3 columns: "Gene"
for gene encoding an enzyme, "Reaction.EC" for categorization of catalytic reactions,
and "Sugar" for degradation pathway. Column names need to be EXACTLY the same,
case sensitive.}
}
\value{
ECCount An integer. Number of different sugar specific enzymes present in the geneVec.
}
\description{
Given a vector of gene names that contain all sugar degradation enzymes from
a genome of interest, count of number of different enzymes encoded by these genes
that involve in the degradation of the specified sugar. Enzymes with same catalytic
function (Reaction.EC) but encoded by different genes are counted once.
}
\examples{
\dontrun{
 geneVec <- c("rpe", "rpiB", "eno", "fruK")
 ER <- EnzymaticReactions
 ECCount <- calculateStepsForOneSugar(geneVec, "fructose", ER)
 ECCount
}

}
