% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competeMicrobiota.R
\name{calculateTotalSteps}
\alias{calculateTotalSteps}
\title{calculate total steps involved in each sugar degradation pathway}
\usage{
calculateTotalSteps(all_sugars, ER)
}
\arguments{
\item{all_sugars}{A list of sugar names whose pathway enzyme count are to
be calculated.}

\item{ER}{A dataset with at least two required columns with specified names.
Sugar (sugar pathway), Reaction.EC (enzymatic reaction type)}
}
\value{
A vector. Name of each value is the sugar degraded, and the value
  is the number of steps involved.
}
\description{
For each sugar, count total number of enzymatic steps required to construct
the entire sugar degradation pathway. Total counts are calculated based on
number of enzymes with different EC number.
}
