% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competeMicrobiota.R
\name{calculateTotalStepsForAllSugars}
\alias{calculateTotalStepsForAllSugars}
\title{Calculate Total Steps In The Degradation Pathway Of All Sugars}
\usage{
calculateTotalStepsForAllSugars(allSugars, ER)
}
\arguments{
\item{allSugars}{A vector of all sugar whose degradation steps are to be calculated.}

\item{ER}{A data.frame with at least two required columns with specified names.
Sugar (sugar pathway), Reaction.EC (enzymatic reaction type)}
}
\value{
A vector summarizing enzymatic steps in each sugar degradation pathway. Counts
    as elements and sugar pathway as names.
}
\description{
For each sugar degradation pathway, count total number of enzymatic steps required.
Total counts are calculated based on number of enzymes with different EC number, not
genes with different names.
}
\examples{
\dontrun{
 library("microCompet")
 ER <- microCompet::EnzymaticReactions
 allSugars <- ER$Sugar
 totalSteps <- calculateTotalStepsForAllSugars(allSugars, ER)
 totalSteps
}

}
