% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInputDataframe.R
\name{checkED}
\alias{checkED}
\title{Check Whether The User Input ED Dataframe Satisfies Function Requirement}
\usage{
checkED(ED, requiredColsVec, firstMicrobe, lastMicrobe)
}
\description{
Check Whether The User Input ED Dataframe Satisfies Function Requirement
}
