% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkUserED.R
\name{checkUserED}
\alias{checkUserED}
\title{Check Whether The User Input ED Dataframe Satisfies Function Requirement}
\usage{
checkUserED(ED, firstMicrobe, lastMicrobe)
}
\arguments{
\item{ED}{The data.frame to be passed in as ED arguments in other functions.}

\item{firstMicrobe}{The value to be passed in as firstMicrobe in other functions.}

\item{lastMicrobe}{The value to be passed in as lastMicrobe in other functions.}
}
\value{
A two-element logical vector. The first indicating whether lastMicrobe needs to
    be updated in respective function, while the second indicating whether the ED frame
    contains unexpected values (other than 0, 1, T/F may be accepted but not recommended)
    in columns between firstMicrobe and lastMicrobe.
}
\description{
Given a data.frame ED and two user specified indices, check whether these arguments can
be used successfully by functions \code{overallSimilarity} and \code{competeMicrobiota}.
This function only check for common requirements, and each function carries out the
data.frame update and function-specific checks.
Function stops early if it identifies errors cannot be corrected by function.
}
\examples{
 \dontrun{
  reportVec <- checkUserED(EnzymeDistribution, 4, 10)
  reportVec
 }

}
