% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overallSimilarity.R
\name{compareTwoGenomes}
\alias{compareTwoGenomes}
\title{Compare Two Genomes And Count Number Of Shared Genes}
\usage{
compareTwoGenomes(genome1, genome2, totalGeneCount)
}
\arguments{
\item{genome1}{A vector of 0 and 1, and each element is named by the gene whose
status it represents. 0 for absence, and 1 for presence.}

\item{genome2}{Same as genome1, but represents another genome}

\item{totalGeneCount}{Total number of genes of interest to be compared.}
}
\value{
An integer indicating number of genes shared by genome1 and genome2.
}
\description{
Given two genomes represented by vectors of 0 and 1 (genes in the same order),
count total number of genes with shared status. This includes genes that are
present in both or neither of the two genomes.
This is a helper only called internally by overallSimilarity.
}
\examples{
\dontrun{
 ED <- microCompet::EnzymeDistribution
 allGenes <-ED$Gene
 geneVec1 <- ED$Gene[4:25]
 genome1 <- transformToVector(geneVec1, allGenes)
 geneVec2 <- ED$Gene[10:35]
 genome2 <- transformToVector(geneVec2, allGenes)
 totalGeneCount <- length(allGenes)
 score <- compareTwoGenomes(genome1, genome2, totalGeneCount)
 score
}

}
