% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competeMicrobiota.R
\name{competeMicrobiota}
\alias{competeMicrobiota}
\title{Visualize the competition by a radarchart.}
\usage{
competeMicrobiota(genome_name, gene_lst)
}
\arguments{
\item{genome_name}{}

\item{gene_lst}{}
}
\description{
Visualize the competiion between a given microbe and other 9 available
microbiota members in the EnzymeDistribution dataset.
}
