% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competeMicrobiota.R
\name{completenessForAllPathways}
\alias{completenessForAllPathways}
\title{Evaluate The Completeness Of All Sugar Degradation Pathways}
\usage{
completenessForAllPathways(geneVec, allSugars, ER, totalSteps)
}
\arguments{
\item{geneVec}{A vector of gene names, representing all sugar degradation enzymes
present in one genome of interest.}

\item{allSugars}{A vector of all sugar degradation pathways.}

\item{ER}{A data.frame describing enzymatic reactions with at least 3 columns: "Gene"
for gene encoding an enzyme, "Reaction.EC" for categorization of catalytic reactions,
and "Sugar" for degradation pathway. Column names need to be EXACTLY the same,
case sensitive.}

\item{totalSteps}{A vector containing counts of total steps for all sugar degradation
pathways. Result of helper function \code{calculateTotalStepsForAllSugars}}
}
\value{
A named vector containing completeness score for all sugar degradation pathways.
}
\description{
Given a genome of interest, described by genomeName and a vector of sugar degradation
genes, evaluate the completeness of all sugar degradation pathways. The completeness
is evaluated by (steps can be catalyzed) over (total steps within a pathway), and this
score should always be between 0 and 1 (inclusive).
}
