% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructFullNetwork.R
\name{constructFullNetwork}
\alias{constructFullNetwork}
\title{Build A Full Simple Sugar Degradation Network For One Genome}
\usage{
constructFullNetwork(genomeName, geneVec, ER)
}
\arguments{
\item{genomeName}{Name of the microbe to be plotted.}

\item{geneVec}{A vector containing gene names of all sugar degradation enzymes present
in the given microbial genome (\code{genomeName}).}

\item{ER}{A data.frame containing all enzymatic reactions of interest, and should represent
enzymatic reactions with at least four characteristics: Gene, Substrate, Product, Sugar.
Header should be EXACTLY the same. (See the sample dataset EnzymaticReactions for example
by \code{?EnzymaticReactions}). Rows containing NA will be removed.}
}
\description{
Given a name of a microbe, a vector of gene names for sugar degradation enzymes,
and a data.frame containing all enzymatic reactions of interest, this function
constructs a network plot that represents all sugar degradation steps the given
microbe/genome can perform.
}
\details{
In the output plot, one vertex represents one compound, and its size suggests the
number of sugar degradation pathways it involves in. You can interpret vertex size
as the importance of a compound within this network.

Similarly, one edge represents one enzymatic reaction that convert the tail vertex (a
reaction substrate) to the head vertex (a reaction product). Thickness of one edge
represents number of genes that encoding enzymes capable of catalyzing this reaction,
and represents the robustness of this step.

Color of the edge represents the sugar degradation pathway this step involved in.
}
\examples{
\dontrun{
 require("microCompet")
 ER <- microCompet::EnzymaticReactions
 ED <- microCompet::EnzymeDistribution
 fullEnzymeGeneVec <- ED$Gene
 genomeFilePath <- system.file("extdata",
                               "Lactobacillus johnsonii.gb",
                               package = "microCompet",
                               mustWork = TRUE)
 carboGenes <- extractCarboGenes(genomeFilePath, fullEnzymeGeneVec)
 fullNetwork <- constructFullNetwork("Lactobacillus johnsonii", carboGenes, ER)
 fullNetwork
}

}
