% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructFullPathway.R
\name{constructFullPathway}
\alias{constructFullPathway}
\title{Build a full sugar degradation pathway}
\usage{
constructFullPathway(gene_lst)
}
\description{
Given an enzyme that involve in sugar degradation, construct the
full pathway in which it participates
}
