% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructFullNetwork.R
\name{createEdgeFrame}
\alias{createEdgeFrame}
\title{Create A Data.frame To Represent All Edges in fullNetwork}
\usage{
createEdgeFrame(relevantReactions)
}
\arguments{
\item{relevantReactions}{A data.frame of enzymatic reactions, containing the required columns:
Gene, Substrate, Product, Sugar (case sensitive).}
}
\value{
A data.frame with 4 columns. Substrate and Product are tail and head of each edge;
    NumOfEnzymes is count of enzymes that can catalyze the reaction; Sugar is the simple sugar
    degradation pathway this reaction is one step of.
}
\description{
Given a list of enzymatic reactions, extract unique reactions, count number of genes encoding
enzymes that catalyze each reaction.
This is a helper function defined for constructFullNetwork and is called only internally.
}
\examples{
\dontrun{
 library("microCompet")
 ER <- microCompet::EnzymaticReactions
 relevantReactions <- ER[ER$Sugar == "ribose", ]
 edgeFrame <- createEdgeFrame(relevantReactions)
 edgeFrame
}

}
