% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructFullNetwork.R
\name{createNodeFrame}
\alias{createNodeFrame}
\title{Create A Data.frame To Represent All Vertices in fullNetwork.}
\usage{
createNodeFrame(edgeFrame)
}
\arguments{
\item{edgeFrame}{A data frame created by createEdgeFrame, must have
"Substrate" and "Product" columns, case sensitive.}
}
\description{
Given an edge frame (have columns "Substrate" and "Product), count total
unique compounds involved in all reactions, and their weight (how many
reactions this compound is involved in.)
This is a helper function defined for constructFullNetwork and is called only internally.
}
\examples{
\dontrun{
 ER <- microCompet::EnzymaticReactions
 relevantReactions <- ER[ER$Sugar == "ribose", ]
 edgeFrame <- createEdgeFrame(relevantReactions)
 nodeFrame <- createNodeFrame(edgeFrame)
 nodeFrame
}

}
