% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EnzymaticReactions}
\alias{EnzymaticReactions}
\title{Enzymatic Reactions Of Simple Sugar Degradation}
\format{
A matrix with 96 rows and 6 variables:
\describe{
 \item{Gene}{Name of the gene that encodes the enzyme represented in this row.}
 \item{Reaction.EC}{Enzyme Commission (EC) number of this enzyme, assigned based
   on the reaction it catalyzes.}
 \item{Enzyme}{Full name of the enzyme.}
 \item{Substrate}{Starting material of the enzyme. If one reaction requires more
   than one substrates, the reaction is split multiple lines.}
 \item{Product}{Product of this enzymatic reaction. If more than one products are
   produced, they are split into multiple lines.}
 \item{Sugar}{The sugar degradation pathway this enzyme participates in.}
}
}
\source{
MetaCyc from metacyc.org
}
\usage{
EnzymaticReactions
}
\description{
A list of 72 enzymatic reactions that involve in microbial degradation of
several simple sugars. Because some reactions can be catalyzed by more than one
enzymes, the dataset have more than 72 rows.
}
\examples{
 \dontrun{
  head(EnzymaticReactions)
}

}
\keyword{datasets}
