% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EnzymeDistribution}
\alias{EnzymeDistribution}
\title{Distribution of Sugar Degradation Enzymes}
\format{
A matrix with 71 rows (genes) and 13 variables
 \describe{
  \item{Gene}{Name of gene encoding this sugar degradation enzyme.}
  \item{Reaction.EC}{EC category of this enzyme, indicating its reaction type.}
  \item{Enzyme}{Full name of the enzyme}
  \item{Sugar}{The sugar degradation pathway this enzyme invoved in.}
  \item{Lplan}{Enzyme distribution in Lactobacillus plantarum}
  \item{Ecoli}{Enzyme distribution in Escherichia coli}
  \item{Blong}{Enzyme distribution in Bifidobacterium longum}
  \item{Paeru}{Enzyme distribution in Pseudomonas aeruginosa}
  \item{Bthet}{Enzyme distribution in Bacteroides thetaiotaomicron}
  \item{Cbotu}{Enzyme distribution in Clostridium botulinum}
  \item{Enter}{Enzyme distribution in Enterobacter sp. EA-01}
  \item{Kvari}{Enzyme distribution in Klebsiella variicola}
  \item{Spneu}{Enzyme distribution in Streptococcus pneumoniae}
 }
}
\source{
Genomes downloaded from NCBI
}
\usage{
EnzymeDistribution
}
\description{
The distribution of 71 sugar degradation enzymes in 9 representative
human microbiota members. "0" suggests the enzyme represented by the
row is not found in the column microbe, while "1" suggests it's found.
}
\details{
E. coli is usually treated as a positive control species since it contains
very abundant metabolic genes and can survive with minimal minerals and
simple sugars.
}
\keyword{datasets}
