% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructFullPathway.R
\name{find_next_gene}
\alias{find_next_gene}
\title{Find the next gene.}
\usage{
find_next_gene(gene_name, direction, all_genes)
}
\description{
Find the next gene within the pathway following the specified direction.
}
