% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GenomesInfo}
\alias{GenomesInfo}
\title{Information about genomes used in EnzymeDistribution}
\format{
An object of class \code{data.frame} with 9 rows and 3 columns.
}
\usage{
GenomesInfo
}
\description{
@format A matrix with 9 rows and 3 variables
 \describe{
  \item{Bacterial.Genome}{Full name of 9 bacteria.}
  \item{Accession}{Accession of this genome on NCBI.}
  \item{Description}{Full description of this genome on NCBI, including strain name.}
  }
}
\keyword{datasets}
