% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competeMicrobiota.R
\name{pathCompleteness}
\alias{pathCompleteness}
\title{Evaluate completeness of sugar degradation paths}
\usage{
pathCompleteness(
  genome_name,
  gene_lst,
  all_sugars,
  ER,
  ED,
  first_microbe,
  last_microbe
)
}
\arguments{
\item{genome_name}{Name of the species of interest}

\item{gene_lst}{A list of gene name that represent the species}

\item{all_sugars}{Name of all sugar pathways relevant}

\item{ER}{A dataset with at least 3 columns. Gene name for an enzyme, Reaction.EC for
one catalytic reaction, and Sugar for degradation pathway. Column names are case
sensitive.}

\item{ED}{A dataset with at least Gene, Reaction.EC, and columns for genome
sugar pathways data from different genomes. See EnzymeDistribution for example.}

\item{first_microbe}{Column index of first microbe genome in dataset ED}

\item{last_microbe}{Column index of last microbe genome in dataset ED. Default set
to the last column of ED.}
}
\description{
Evalute path completeness of all sugar degradation pathways listed in all_sugars
}
