% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMicroCompet.R
\name{runMicroCompet}
\alias{runMicroCompet}
\title{Launch Shiny App For microCompet}
\usage{
runMicroCompet()
}
\value{
No values return, this function opens a Shiny page.
}
\description{
This function launches the Shiny app for this package microCompet, which
targeted users with limited background in R language.
}
\examples{
 \dontrun{
  microCompet::runMicroCompet()
 }

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
