% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleGenomeKvari}
\alias{sampleGenomeKvari}
\title{Sample carboGenes extracted from Klebsiella variicola}
\format{
An object of class \code{data.frame} with 40 rows and 1 columns.
}
\usage{
sampleGenomeKvari
}
\description{
A list of all simple sugar degradation enzymes present in Klebsiella variicola.
}
\examples{
\dontrun{
sampleGenomeKvari
}
}
\keyword{datasets}
