% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleGenomeLjohn}
\alias{sampleGenomeLjohn}
\title{Sample carboGenes extracted from Lactobacillus johnsonii}
\format{
An object of class \code{data.frame} with 13 rows and 1 columns.
}
\usage{
sampleGenomeLjohn
}
\description{
A list of all simple sugar degradation enzymes present in Lactobacillus johnsonii.
}
\examples{
\dontrun{
sampleGenomeLjohn
}
}
\keyword{datasets}
