% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overallSimilarity.R
\name{transformToVector}
\alias{transformToVector}
\title{Transform a list of gene into a vector of 0 and 1}
\usage{
transformToVector(gene_lst, all_genes)
}
\arguments{
\item{gene_lst}{A list of gene_names, represented by 3-5 characters, such as "eno".}

\item{all_genes}{A list of all sugar degradation genes that may be possibily included.}
}
\value{
A vector of 0 and 1, indicating whether each sugar degradation enzyme (from
the EnzymeDistribution dataset) is included in the gene_lst.
}
\description{
Given a list of genes, transform it into a vector of 0 and 1.
Each cell is a named gene, 0 indicating this genome is not included
inside the gene_lst, 1 indicating it's included
}
\examples{
\dontrun{
 ED <- microCompet::EnzymeDistribution
 all_genes <- ED$Gene
 gene_lst <- ED$Gene[4:25]
 genome_vector <- transformToVector(gene_lst, all_genes)
 genome_vector
}

}
