% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competeMicrobiota.R
\name{allSugarScoresForOneGenome}
\alias{allSugarScoresForOneGenome}
\title{Calculate pathway completeness for all pathways within a genome.}
\usage{
allSugarScoresForOneGenome(geneVec, allSugars, ER)
}
\arguments{
\item{geneVec}{Gene list represents on genome}

\item{allSugars}{A list representing all sugar degradation pathways.}

\item{ER}{A dataset with at least 3 columns. Gene name for an enzyme, Reaction.EC for
one catalytic reaction, and Sugar for degradation pathway. Same requirement for
the next calculateStepsForOneSugar function, since it's only for passing to the helper.}
}
\value{
A vector containing a genome's score for all sugar degradatin pathways.
}
\description{
Use the calculateStepsForOneSugar function for all sugar pathways one by one,
and build a named vector for all pathways
}
\examples{
\dontrun{
 geneVec <- c("rpe", "rpiB", "eno", "fucK")
 ER <- microCompet::EnzymaticReactions
 ED <- microCompet::EnzymeDistribution
 allSugars <- sort(unique(ED$Sugar))
 ECCounts <- allSugarScoresForOneGenome(geneVec, allSugars, ER)
 ECCounts
}

}
