% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competeMicrobiota.R
\name{allSugarScoresForOneGenome}
\alias{allSugarScoresForOneGenome}
\title{Calculate pathway completeness for all pathways within a genome.}
\usage{
allSugarScoresForOneGenome(gene_lst, all_sugars, ER)
}
\arguments{
\item{gene_lst}{Gene list represents on genome}

\item{all_sugars}{A list representing all sugar degradation pathways.}

\item{ER}{A dataset with at least 3 columns. Gene name for an enzyme, Reaction.EC for
one catalytic reaction, and Sugar for degradation pathway. Same requirement for
the next calculateCount function, since it's only for passing to the helper.}
}
\value{
A vector containing a genome's score for all sugar degradatin pathways.
}
\description{
use the calculateCount function for all sugar pathways.
}
\examples{
\dontrun{
 gene_lst <- c("rpe", "rpiB", "eno", "fucK")
 ER <- microCompet::EnzymaticReactions
 ED <- microCompet::EnzymeDistribution
 all_sugars <- sort(unique(ED$Sugar))
 enzyme_count <- calculateCount(gene_lst, all_sugars, ER)
 enzyme_count
}


}
