% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competeMicrobiota.R
\name{calculateCount}
\alias{calculateCount}
\title{Calculate number of sugar-degradation specific enzymes in gene_lst.}
\usage{
calculateCount(gene_lst, sugar)
}
\arguments{
\item{gene_lst}{A list of gene names represented by 3-5 letters, such as "tktA".}

\item{sugar}{The desired sugar degradation pathway. A simple sugar such as "fructose".}
}
\value{
enzyme_count Number of sugar_specific enzymes inside the gene_lst.
}
\description{
Given a gene_lst, probably represent all sugar-degradatinon enzymes from a genome,
extract enzymes encoded by these genes that involve in the specific degradation
pathway of the given sugar. Enzymes with same catalytic function but encoded by
different genes are treated as the same.
}
