% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overallSimilarity.R
\name{compareTwoGenomes}
\alias{compareTwoGenomes}
\title{Compare two genomes and count shared genes}
\usage{
compareTwoGenomes(genome1, genome2, total_genes)
}
\arguments{
\item{genome1}{A vector of 0 and 1, genes ordered as in EnzymaticDistribution}

\item{genome2}{Same as genome1, but represents another genome}

\item{total_genes}{Total number of genes of interest to be compared.}
}
\value{
An integer indicating number of shared genes by genome1 and genome2.
}
\description{
Given two genomes represented by a vector of 0 and 1 (genes in the same order),
count shared genes. This includes genes both genomes have or neither.
}
