% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competeMicrobiota.R
\name{competeMicrobiota}
\alias{competeMicrobiota}
\title{Visualize the competition by a radarchart.}
\usage{
competeMicrobiota(
  genome_name,
  gene_lst,
  ER,
  ED,
  first_microbe,
  last_microbe = ncol(ED)
)
}
\arguments{
\item{genome_name}{Steing represents the name of a genome/microbe.}

\item{gene_lst}{A list of sugar degradation gene names in microbe genome_name}

\item{ER}{A dataset with at least 3 columns. Gene name for an enzyme, Reaction.EC for
one catalytic reaction, and Sugar for degradation pathway. Column names are case
sensitive.}

\item{ED}{A dataset with at least Gene, Reaction.EC, and columns for genome
sugar pathways data from different genomes. See EnzymeDistribution for example.}

\item{first_microbe}{Column index of first microbe genome in dataset ED}

\item{last_microbe}{Column index of last microbe genome in dataset ED. Default set
to the last column of ED.}
}
\description{
Visualize the competiion between a given microbe and other 9 available
microbiota members in the EnzymeDistribution dataset.
}
\examples{
 \dontrun{
 library(microCompet)
 genome_name <- "L. johnsonii"
 ED <- microCompet::EnzymeDistribution
 full_enzyme_gene_lst <- ED$Gene
 genome_file_path = "./Lactobacillus_johnsonii.gb"
 carbo_genes <- extractCarboGenes(genome_file_path, full_enzyme_gene_lst)
 first_microbe = 5
 last_microbe = 13
 ER <- microCompet::EnzymaticReactions
 compete_microbiota <- competeMicrobiota(genome_name, carbo_genes, ER,
                                         ED, first_microbe, last_microbe)
 compete_microbiota
 }

}
