% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructFullNetwork.R
\name{constructFullNetwork}
\alias{constructFullNetwork}
\title{Build a full sugar degradation network for one genome.}
\format{
A dataframe with four required columns, can have other columns.
 \describe{
  \item{Gene}{Genes encoding the enzyme}
  \item{Substrate}{Substrate of the enzyatic reaction.}
  \item{Product}{Product of the reaction.}
  \item{Sugar}{Sugar degradation pathway this enzyme participates in.}
 }
}
\usage{
constructFullNetwork(genome_name, gene_lst, ER)
}
\arguments{
\item{genome_name}{The name of microbe to be graphed.}

\item{gene_lst}{A list of gene names that represent sugar degradation enzymes.}

\item{ER}{A list of enzymatic_reactions, containing only enzymes
encoded by genes from gene_lst.}
}
\description{
Given a list of sugar degradation enzymes, construct a metabolism network
of these enzymes and create a network plot.
One vertex represents one compound. Size of the vertex suggests its importance,
the larger the vertex, the more pathways this compound is involved in.
One edge represents one enzymatic reaction that convert the tail vertex (a
compound) to the head vertex (another compound). The thicker the edge, the more
enzymes can catalyze this reaction.
Color of the edge represents the sugar degradation pathway this enzyme involved in.
}
\examples{
\dontrun{
 require("microCompet")
 ER <- microCompet::EnzymaticReactions
 ED <- microCompet::EnzymeDistribution
 full_enzyme_gene_lst <- ED$Gene
 genome_file_path <- "./Klebsiella_variicola.gb"
 carbo_genes <- extractCarboGenes(genome_file_path, full_enzyme_gene_lst)
 full_pathway <- constructFullNetwork("Kvari", carbo_genes, ER)
 full_pathway
}

}
