% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructFullNetwork.R
\name{constructFullNetwork}
\alias{constructFullNetwork}
\title{Build a full sugar degradation network for one genome.}
\usage{
constructFullNetwork(gene_lst)
}
\arguments{
\item{gene_lst}{A list of gene names that represent sugar degradation enzymes.}
}
\description{
Given a list of sugar degradation enzymes, construct a metabolism network
of these enzymes and create a network plot.
One vertex represents one compound. Size of the vertex suggests its importance,
the larger the vertex, the more pathways this compound is involved in.
One edge represents one enzymatic reaction that convert the tail vertex (a
compound) to the head vertex (another compound). The thicker the edge, the more
enzymes can catalyze this reaction.
Color of the edge represents the sugar degradation pathway this enzyme involved in.
}
