% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructFullNetwork.R
\name{createEdgeFrame}
\alias{createEdgeFrame}
\title{Create a dataframe to be used as edges for FullNetwork}
\format{
A dataframe with four required columns, can have other columns.
 \describe{
  \item{Gene}{Genes encoding the enzyme, though not directly accessed, it's counted}
  \item{Substrate}{Substrate of the enzyatic reaction.}
  \item{Product}{Product of the reaction.}
  \item{Sugar}{Sugar degradation pathway this enzyme participates in.}
 }

\describe{
  \item{substrate}{substrate of this reaction, tail of the edge}
  \item{product}{product of the reaction, head of the edge}
  \item{num_of_enzymes}{numbef of genes encoding enzymes that can catalyze this reaction}
  \item{sugar}{the sugar degradation pathway this enzyme participates in.}
 }
}
\usage{
createEdgeFrame(relevant_reactions)
}
\arguments{
\item{relevant_reactions}{A list of enzymatic_reactions, containing only enzymes
encoded by genes from gene_lst.}
}
\value{
A dataframe with 4 columns
}
\description{
Given a list of enzymatic reactions, extract unique reactions, count genes encoding
enzymes for each reaction.
This function is only called as a helper for constructFullNetwork.
}
\examples{
\dontrun{
 library("microCompet")
 ER <- microCompet::EnzymaticReactions
 relevant_reactions <- ER[ER$Sugar == "ribose", ]
 edge_frame <- createEdgeFrame(relevant_reactions)
 edge_frame
}

}
