% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructFullNetwork.R
\name{createEdgeFrame}
\alias{createEdgeFrame}
\title{Create a dataframe to be used as edges for FullNet}
\format{
\describe{
  \item{substrate}{substrate of this reaction, tail of the edge}
  \item{product}{product of the reaction, head of the edge}
  \item{num_of_enzymes}{numbef of genes encoding enzymes that can catalyze this reaction}
  \item{sugar}{the sugar degradation pathway this enzyme participates in.}
 }
}
\usage{
createEdgeFrame(relevant_reactions)
}
\arguments{
\item{relevant_reactions}{A list of enzymatic_reactions, containing only enzymes
encoded by genes from gene_lst (input of constructFullNet).}
}
\value{
A dataframe with 4 columns
}
\description{
Given a list of enzymatic reactions (a subset of the EnzymaticReactions dataset),
extract unique reactions, count genes encoding enzymes for each reaction, and
keep the "sugar" column for coloring purpose.
This function is only called as a helper for constructFullNet.
}
