% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructFullNetwork.R
\name{createNodeFrame}
\alias{createNodeFrame}
\title{Create a dataframe to be used as vertices for FullNetwork.}
\format{
\describe{
  \item{compound}{the compound this vertex represents.}
  \item{weight}{number of edges this vertex involves in}
 }
}
\usage{
createNodeFrame(edge_frame)
}
\arguments{
\item{edge_frame}{A data frame created by createEdgeFrame, must have
"substrate", "product" columns, case sensitive.}
}
\value{
A dataframe with 2 columns.
}
\description{
Given an edge frame (have columns "substrate" and "product), count total
unique compounds involved in all reactions, and their weight (how many
reactions this compound is involved in.)
This function is only called as a helper for constructFullNetwork.
}
