% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructFullNetwork.R
\name{createNodeFrame}
\alias{createNodeFrame}
\title{Create a dataframe to be used as edges for FullNet}
\format{
\describe{
  \item{compound}{the compound this vertex represents.}
  \item{weight}{number of edges this vertex involves in}
 }
}
\usage{
createNodeFrame(edge_frame)
}
\arguments{
\item{relevant_reactions}{A list of enzymatic_reactions, containing only enzymes
encoded by genes from gene_lst (input of constructFullNetwork).}
}
\value{
A dataframe with 2 columns
}
\description{
Given a list of enzymatic reactions (a subset of the EnzymaticReactions dataset),
extract unique reactions, count genes encoding enzymes for each reaction, and
keep the "sugar" column for coloring purpose.
This function is only called as a helper for constructFullNetwork.
}
