% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCarboGenes.R
\name{extractCarboGenes}
\alias{extractCarboGenes}
\title{Extract Carbohydrate Degradation Genes from Microbe Genome}
\usage{
extractCarboGenes(genomeFilePath, fullEnzymeGeneList)
}
\arguments{
\item{genomeFilePath}{Full path to a GenBank file that represents a microbial genome.
This GenBank file must at least have gene names annotated, see provided file
"Klebsiella variicola.gb" and "Lactobacillus johnsonii.gb" in extdata for example,
they have lines in the format /gene="gene_name".}

\item{fullEnzymeGeneList}{A list of all sugar degradation genes to be searched for in
the user-provided genome (the file with genomeFilePath).}
}
\value{
A list of gene names that encode simple sugar degradation enzymes.
}
\description{
This function reads a genbank file specified by path line by line and
extracts genes encoding enzymes that participate in simple sugar
degradation from the given genome.
}
\details{
There are other available packages that format the whole genbank file into
one organized object, but information other than geneName are not
useful to this package, and those functions would take about 30 more
seconds on a Windows system, so this function is provided.
}
\examples{
\dontrun{
 require("microCompet")
 ED <- microCompet::EnzymeDistribution
 fullEnzymeGeneList <- ED$Gene
 genomeFilePath <- system.file("extdata",
                               "Klebsiella variicola.gb",
                               package = "microCompet",
                               mustWork = TRUE)
 carboGenes <- extractCarboGenes(genomeFilePath, fullEnzymeGeneList)
 carboGenes
 # carboGenes should be the same as the following geneList
 expectedCarboGenes <- microCompet::sampleGenomeKvari
}

}
