% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCarboGenes.R
\name{extractCarboGenes}
\alias{extractCarboGenes}
\title{Extract Carbohydrate-degrading Genes from Microbe Genome}
\usage{
extractCarboGenes(genome_file_path)
}
\arguments{
\item{genome_file_path}{Full path to a GenBank file that represents a microbial genome.
This GenBank file must at least have gene names annotated.
Only enzymes listed in the EnzymaticReactions dataset will be extracted.}
}
\value{
A list of gene names that encode simple sugar degradation enzymes.
}
\description{
This functions extract genes encoding enzymes that participate in simple
sugar degradation from the given genome.
}
\examples{
\dontrun{
 file_path = "Path_to_File.gb"
 carboGenes <- extractCarboGenes(file_path)
}

}
