% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GenomesInfo}
\alias{GenomesInfo}
\title{Information About Genomes Included In EnzymeDistribution}
\format{
A matrix with 9 rows and 3 variables
 \describe{
  \item{Bacterial.Genome}{Full name of bacterial species.}
  \item{Accession}{Accession of this genome on NCBI.}
  \item{Description}{Full description of this genome on NCBI, including strain name.}
}
}
\usage{
GenomesInfo
}
\description{
This dataset offers detailed information about genomes included in another dataset
EnzymeDistribution (See with \code{?EnzymeDistribution}). You can find the original
genbank files on NCBI by searching the Accession.
}
\examples{
\dontrun{
  GenomesInfo
}

}
\keyword{datasets}
