% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overallSimilarity.R
\name{overallSimilarity}
\alias{overallSimilarity}
\title{Compare overall similarity of the given genomes and microbiota.}
\usage{
overallSimilarity(genome_name, gene_lst)
}
\arguments{
\item{genome_name}{Name of the genome represented by gene_lst}

\item{gene_lst}{A list of gene names represented by 3-5 characters, such as "rpiB"}
}
\value{
A radar chart with microbial species on corners and the user input genome
   in the middle.
}
\description{
Based on sugar degradation genes, compare the overall similarity
of the given genome and other 9 microbial species from microbiota
available species in the EnzymeDistribution dataset.
}
