% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overallSimilarity.R
\name{overallSimilarity}
\alias{overallSimilarity}
\title{Compare Overall Similarity Of A Genome And Microbiota Members}
\usage{
overallSimilarity(
  genomeName,
  geneVec,
  ED,
  firstMicrobe,
  lastMicrobe = ncol(ED)
)
}
\arguments{
\item{genomeName}{Name of the genome represented by geneVec}

\item{geneVec}{A vector gene names represented by 3-5 characters, such as "rpiB",
containing all present simple sugar degradation enzymes in the microbe with
genomeName.}

\item{ED}{An enzyme distribution data.frame that represents sugar degradation enzyme
profiles in genomes that genomeName microbe is to be compare with. This data.frame
contains a column "Gene" (case sensitive), and at least one genome profile
from column index firstMicrobe to column index lastMicrobe CONTINUOUSLY. Genome
profiles use 1 to indicate the presence of a gene, and 0 for absence. T/TRUE and
F/FALSE can be coerced into 1 and 0, but not recommended.}

\item{firstMicrobe}{A positive integer. Index of the first column in ED data.frame that
represents a microbial genome. Count starts from 1. That is if ED contains microbial
genome profiles from column 4-9, firstMicrobe is 4.}

\item{lastMicrobe}{A positive integer. Index of the last column in ED data.frame that
represents a microbial genome. Count starts from 1. That is if ED contains microbial
genome profiles from column 4-9, firstMicrobe is 9. Default of lastMicrobe is
the last column of data.frame ED.}
}
\value{
A radar chart with microbial species on corners and the user input genome
    in the middle. The genome of interest (genomeName) receives a score when compred
    with each genome in the data.frame ED, and is formatted as [number of shared genes]
    out of [total number of genes of interest].
}
\description{
Based on the presence of simple sugar degradation enzymes, this function compares
the overall similarity of one genome with other genome profiles provided as a
data.frame.
}
\examples{
\dontrun{
 require("microCompet")
 genomeName <- "L. johnsonii"
 ED <- microCompet::EnzymeDistribution
 fullEnzymeGeneVec <- ED$Gene
 genomeFilePath <- system.file("extdata",
                               "Lactobacillus johnsonii.gb",
                               package = "microCompet",
                               mustWork = TRUE)
 carboGenes <- extractCarboGenes(genomeFilePath, fullEnzymeGeneVec)
 overSimi <- overallSimilarity(genomeName, carboGenes, ED, 5, 13)
 overSimi
}

}
