% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMicroCompet.R
\name{runMicroCompet}
\alias{runMicroCompet}
\title{Launch Shiny App For microCompet}
\usage{
runMicroCompet()
}
\value{
No values return, this function opens a Shiny page.
}
\description{
-------------.
}
\examples{
 \dontrun{
  microCompet::runMicroCompet()
 }

}
