% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overallSimilarity.R
\name{transformToVector}
\alias{transformToVector}
\title{Transform A Vector Of Gene Names Into A Vector Of 0 And 1}
\usage{
transformToVector(geneVec, allGenes)
}
\arguments{
\item{geneVec}{A vector of gene names, represented by 3-5 characters, such as "eno",
this vector includes all simple sugar degradation enzymes present in a microbial
genome of interest.}

\item{allGenes}{A vector containing all sugar degradation genes of interest.}
}
\value{
A named vector of 0 and 1, indicating whether each sugar degradation enzyme,
    represented by the element's name, is presetn in the geneVec.
}
\description{
Given a vector of genes that represents a genome, transform it into a simple sugar
degradation enzyme profile as a vector of 0 and 1. A 0 indicates the gene represented
by element name is absent in the genome, and a 1 indicates its presence.
}
\examples{
\dontrun{
 ED <- microCompet::EnzymeDistribution
 allGenes <- ED$Gene
 geneVec <- ED$Gene[4:25]
 genomeVec <- transformToVector(geneVec, allGenes)
 genomeVec
}

}
