% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overallSimilarity.R
\name{transformToVector}
\alias{transformToVector}
\title{Transform a list of gene into a vector of 0 and 1}
\usage{
transformToVector(gene_lst)
}
\arguments{
\item{gene_lst}{A list of gene_names, represented by 3-5 characters, such as "eno".}
}
\value{
A vector of 0 and 1, indicating whether each sugar degradation enzyme (from
the EnzymeDistribution dataset) is included in the gene_lst.
}
\description{
Given a list of genes, transform it into a vector of 0 and 1.
Each cell is a named gene, 0 indicating this genome is not included
inside the gene_lst, 1 indicating it's included
}
